files reorg, add github CI for rustfmt checks
