+++ .github/workflows/rustfmt.yml
@@ -1,0 +1,21 @@
+name: Rustfmt Check
+
+on:
+  pull_request:
+    branches: [ main ]
+    types: [ready_for_review, opened, synchronize, reopened]
+  push:
+    branches: [ main ]
+
+jobs:
+  rustfmt:
+    if: github.event.pull_request.draft == false
+    name: Rust formatting
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v4
+      - uses: actions-rust-lang/setup-rust-toolchain@v1
+        with:
+          components: rustfmt
+      - name: Check formatting
+        uses: actions-rust-lang/rustfmt@v1
