+++ .github/workflows/typos.toml
@@ -1,2 +1,3 @@
 [default.extend-words]
 groth = "groth" # to avoid it dectecting it as 'growth'
+BA = "BA"
