+++ src/backends/mock_signed.rs
@@ -1,11 +1,12 @@
+use anyhow::Result;
+use std::any::Any;
+use std::collections::HashMap;
+
 use crate::middleware::{
     containers::Dictionary, hash_str, AnchoredKey, Hash, Params, Pod, PodId, PodSigner, PodType,
     Statement, Value, KEY_SIGNER, KEY_TYPE,
 };
 use crate::primitives::merkletree::MerkleTree;
-use anyhow::Result;
-use std::any::Any;
-use std::collections::HashMap;
 
 pub struct MockSigner {
     pub pk: String,
